Refactoring get method and create a template handler to list data
