+++ app/src/scripts/controller/UserController.js
@@ -50,52 +50,11 @@ class UserController{
         const userList = JSON.parse(localStorage.getItem('users'));
         const tempUserList = JSON.parse(localStorage.getItem('tempUsers'));
         const userListEl = document.getElementById('list-user');
-        if (userListEl != null && userList != null){
-            userListEl.innerHTML = '';
-
-            userList.forEach((el) => {
-                const listItem = document.createElement('li');
-                listItem.classList.add('list-user-item');
-                listItem.setAttribute('id', el.id);
-                        template = `<div class="user-name-content">
-                                        <span class="list-text" id="user-name">${el.name} - ${el.cpf}</span>
-                                    </div>
-                                    <div class="user-full-content">
-                                        <div class="user-info">
-                                            <span class="list-text" id="user-email"><strong>Email:</strong> ${el.email}</span>
-                                            <span class="list-text" id="user-phone"><strong>Tel:</strong> ${el.phone}</span>
-                                        </div>
-                                        <div class="user-action">
-                                            <button type="button" class="action-btn edit icon-pencil" data-id=${el.id} id="edit"></button>
-                                            <button type="button" class="action-btn delete icon-cancel-circle" data-id=${el.id}></button>
-                                        </div>
-                                    </div>`;
-                listItem.innerHTML = template;
-                userListEl.appendChild(listItem);
-            });
-        } 
-        if (userListEl != null && tempUserList != null && userList != null){
-            if (userList == null || userList.length == 0) {
-                userListEl.innerHTML = '';
-                tempUserList.forEach((el) => {
-                    const listItem = document.createElement('li');
-                    listItem.classList.add('list-user-item');
-                    listItem.setAttribute('id', el.id);
-                    template = `<div class="user-name-content">
-                                            <span class="list-text" id="user-name">${el.name} - ${el.cpf}</span>
-                                        </div>
-                                        <div class="user-full-content">
-                                            <span class="list-text" id="user-email">Email: ${el.email}</span>
-                                            <span class="list-text" id="user-phone">Tel: ${el.phone}</span>
-                                        </div>`;
-                    listItem.innerHTML = template;
-                    userListEl.appendChild(listItem);
-                }) ;              
-                
-            }
+        if (userList != null && userList.length > 0){
+            return userList;
+        }else if(tempUserList != null && tempUserList.length > 0){
+            return tempUserList;
         }
-        
-        
     }
 
     promiseGet(){
